refactor: update factory imports in CRUD templates
