+++ cli/internal/template/crud/create/run.go
@@ -1,6 +1,6 @@
 package create
 
-import "github.com/aerogear/charmil/core/factory"
+import "github.com/aerogear/charmil/cli/internal/factory"
 
 func runCreate(opts *createOptions, f *factory.Factory) error {
 	// Add your implementation here
