+++ cli/internal/template/crud/delete/run.go
@@ -1,6 +1,6 @@
 package delete
 
-import "github.com/aerogear/charmil/core/factory"
+import "github.com/aerogear/charmil/cli/internal/factory"
 
 func runDelete(opts *deleteOptions, f *factory.Factory) error {
 	// Add your implementation here
