+++ cli/internal/template/crud/describe/run.go
@@ -1,6 +1,6 @@
 package describe
 
-import "github.com/aerogear/charmil/core/factory"
+import "github.com/aerogear/charmil/cli/internal/factory"
 
 func runDescribe(opts *describeOptions, f *factory.Factory) error {
 	// Add your implementation here
