+++ cli/internal/template/crud/list/run.go
@@ -1,6 +1,6 @@
 package list
 
-import "github.com/aerogear/charmil/core/factory"
+import "github.com/aerogear/charmil/cli/internal/factory"
 
 func runList(opts *listOptions, f *factory.Factory) error {
 	// Add your implementation here
