+++ cli/internal/template/crud/use/run.go
@@ -1,6 +1,6 @@
 package use
 
-import "github.com/aerogear/charmil/core/factory"
+import "github.com/aerogear/charmil/cli/internal/factory"
 
 func runUse(opts *useOptions, f *factory.Factory) error {
 	// Add your implementation here
